+++ src/Video/VideoThumbGeneratorFactory.php
@@ -12,19 +12,14 @@ declare(strict_types=1);
 namespace Soluble\MediaTools\Video;
 
 use Psr\Container\ContainerInterface;
-use Psr\Log\NullLogger;
 use Soluble\MediaTools\Video\Config\FFMpegConfigInterface;
-use Soluble\MediaTools\Video\Config\LoggerConfigInterface;
+use Soluble\MediaTools\Video\Logger\LoggerInterface;
 
 class VideoThumbGeneratorFactory
 {
     public function __invoke(ContainerInterface $container): VideoThumbGeneratorInterface
     {
-        if ($container->has(LoggerConfigInterface::class)) {
-            $logger = $container->get(LoggerConfigInterface::class)->getLogger();
-        } else {
-            $logger = new NullLogger();
-        }
+        $logger = $container->has(LoggerInterface::class) ? $container->get(LoggerInterface::class) : null;
 
         return new VideoThumbGenerator(
             $container->get(FFMpegConfigInterface::class),