+++ tests/util/ServicesProviderTrait.php
@@ -108,8 +108,9 @@ trait ServicesProviderTrait
 
         $config = [
             'soluble-mediatools' => [
-                'ffmpeg.binary'  => $ffmpegBinary,
-                'ffprobe.binary' => $ffprobeBinary,
+                'ffmpeg.binary'          => $ffmpegBinary,
+                'ffprobe.binary'         => $ffprobeBinary,
+                'logger.service_factory' => \Soluble\MediaTools\Video\Logger\NullLoggerFactory::class,
             ],
         ];